Decouple AuthenticationMethod from protocol implementation
The interface contained separate methods for http and pg - the intention
was to be able to support more complex authentication methods - e.g.
password authentication.

But these will require a more complicated message flow - so we'll have
to re-model the interface and we should try to decouple the state
changes from the protocol implementation.