LICENSE and NOTICE improvements
CrateDB is not a product of the Apache Software Foundation (ASF),
however, we have decided that it is prudent to follow the ASF's own
guidelines for applying the Apache License 2.0 to our project.

We are following most of the principals set out in this document:

    https://www.apache.org/dev/apply-license.html

However, some principals do not apply, or we have chosen not to apply
them on a case-by-case basis.

For the purposes of this patch, the three main changes are:

- Removed several out-of-date sections from LICENSE. Specifically, if we
  are not bundling the bits of a third-party dependency in our Git
  repository, we are not documenting that software for the purposes of
  LICENSE or NOTICE.

- We have made a minor exception to the above principal, and have chosen
  to document the nature of the `es/upstream` Git submodule.

- The NOTICE file has been improved with a few small fixes and a better
  description of how we are handling the documentation of licensing
  matters.