Prepare CrateNettyHttpServerTransport to be de-guiced
This extracts a `PipelineRegistry` from the
`CrateNettyHttpServerTransport` in order to prepare it for the ES 5.1
update.

In ES 5.1 `onModule` has been removed. To register a different http
transport there is a new `NetworkPlugin` interface which contains a
method as follows:

     Map<String, Supplier<HttpServerTransport>> getHttpTransports(...)

This can be used to create and register the
`CrateNettyHttpServerTransport`, but as a side effect it won't be bound
to guice anymore - which means it's no longer a global singleton. This
breaks the pipeline logic.

To work around that there is a new `PipelineRegistry` which can be
created earlier and be made available within guice.