Fix partitioned-by/generated-column COPY FROM IndexOutOfBoundsException
The InputColumn index calculation in the creation of the
SourceIndexWriterProjection wasn't correct.

This commit changes the pattern to create the projection from an
imperative to a more declarative approach.

The input-column indices are no longer calculated manually. Instead the
`toCollect` symbols are declared based on primary keys and partitioned
by columns. Then the `InputColumns` API is used to generate the input
columns.

This also eliminates the loose contract that we had, that the
`SourceIndexWriterProjection` constructor expected the `toCollect`
symbols of the `FileUriCollectPhase` to be in a certain order.