Fixed weird log message coming from the parser.
SymbolPrinter was used directly to build paths for Fields of TwoTableJoin.
This caused the parser to recognize keyword "join" used in the (QualifiedNames
of the joins) and log a message "line 1:0 no viable alternative at input 'join'".
Replacing direct usage of SymbolPrinter with Symbols.pathFromSymbol() solves the issue
as it avoids calling the SymbolPrinter if the symbol is a Field.