Fix insert error handling & fix testDeletePartitionWhileBulkInsertingData
A IndexNotFoundException in the UpsertByIdTask could be swallowed.

In addition this commit changes the tests in
`PartitionedTableConcurrentIntegrationTest` to be less resource intense.
There was a chance that the first requests would run into a timeout when
creating the index + waiting for the shards to become active.

This could even lead to a "shard leak" because the test could finish
with the request still being active.