Fixed TableInfo creation for unnest().
The comparator for sorting the columns numericly and not lexically is also used when issuing
get() on the ImmutableSortedMap/Set so it should be implemented with more complex logic.

Instead of that we skip ColumnRegistrar and we create the Map and Set
of columns presorted as we need.

This is a followup of commit: e1a76e1ea0628d423b9eb09fd80c141f68dc5401