Apply CircuitBreaker logic to PagingIterators.
During a distributed operation the memory used in the PagingIterators
is not accounted at all which leads to OOMemoryExceptions which results
in killed nodes.

Added a new RamAccountingPageIterator which wraps the actual PagingIterator, calculates
the memory used and produces a CircuitBreakingException when the limits are exceeded.
So instead of having OOMemoryExecptions which kill node(s) of the cluster the job is killed
and the cluster is unaffected.