Make sure DELETE statements with _version filter don't fail
This removes the ESJobContext abstraction:

The InternalActionListener called `context.close(e);` - this by-passed
the error handling in the `ESDeleteTask` and caused a race condition, in
that sometimes a VersionConflictEngineException would be propagated to
the user instead of being turned into row-count 0

`testDeleteWhereVersionWithConflict` was flaky because of this.

Having a context for the delete case is mostly just overhead. There is
no way to KILL an ongoing delete request.

This also causes us to loose `sys.operations` entries, but it's already
documented that there may be operations without an entry, and we still
have the `sys.logs` entry.