Create a new request after submitting the `current` one for execution.
For bulk operations (like `delete from...`, `update...where`) we collect the
information about the target items until we reach the bulk size, at which point
we execute the operation for the current bulk. While executing the current bulk
we concurrently collect information about the items in the next bulk.
We "buffer"/"collect" the items in the "next to be executed" bulk in a request
object. Due to the concurrent nature of the collection and execution of bulks
we need to make sure we use a *different* request object to collect the items
for the next bulk (while the current one is executing).

(cherry picked from commit 44c87879cea5b72b577b0c15878cfeb5c682b855)
Signed-off-by: Andrei Dan <andrei@crate.io>