Reduce amount of tests run on travis
The full test suite is taking too long (> 45 min) and is now often
failing because of the time limit on travis.

All tests are also run on Jenkins - so this reduces the tests that are
run on Travis to `:sql:test` - this should still be useful as flaky
tests tend to fail more often on Travis.