Fix issue with JOIN on >=4 tables with ORDER BY.
When building the join tree in the ManyTableConsumer we try to find the optimum order
of relations based also on the ORDER BY. The ORDER BY has the highest priority so we
move the tables that are included in the ORDER BY symbols to the left side of the tree
so that are joined as early as possible. If the join conditions contain relations that
are not immediately adjucent in the original query (eg: ... t2 JOIN t3 on t3.id = t1.d)
the new optimized ordering could lead to a situation that we try to apply a join condition
that contains relations which are not already part of the tree.

To solve this problem a new Map structure is build which holds the join conditions and
helps to apply the following logic:

During the construction of the join tree at the depth that is currently being built with
with the help of the new Map we try to apply all the eligible join conditions for that depth,
so we end up with a join tree where the join conditions are applied as early as possible.

This commit fixes: https://github.com/crate/crate/issues/5747