Add `empty_row` table function and use it implicitly in statements without `FROM`
Statements without `FROM` would implicitly use the `sys.cluster` table
as a target. `SELECT 1` is often used as a connection check and we still
want to support this after the schema privileges feature is rolled out.
This commit replaces the `sys.cluster` relation with the `empty_row` table
function relation that generates an empty row without any column. This
will allow us to use the control `SELECT 1` statement after schema
privileges are relesead (as the user will not need special rights on
the `sys` schema).