Fix false decrement of reference counted messages
The HttpAuthUpstreamHandler grants permission for messages in the
Netty pipeline. It uses the SimpleChannelInboundHandler which, by
default, releases reference counted messages after calling the channel's
read method. This leads to an unwanted decrement of messages and results
in downstream errors like the following:
io.netty.util.IllegalReferenceCountException: refCnt: 0, decrement: 1

This disables automatic releasing and releases the reference counted
message manually only in case it is not forwarded to the downstream
handler.