Re-work user functionality loading mechanism
This replaces the (ES) UsersPlugin with a UsersExtension class that is
loaded using a ServiceLoader from within the SQLPlugin.

This solves a ClassCastException that occured due to the usage of two
independent URLClassLoader instances. (One from the PluginService, one
from the EnterpriseLoader). This resulted in the loss of custom metadata
since the ES52 upgrade on node start, as the metadata couldn't be read
anymore.