Fix issue with count(*) on >= 3 table joins.
When using count(*) and implicit join conditions (in where clause)
then the outputs of the mss.querySpec() is empty. So we need to build
those outputs on every join pair of the join tree (except for the top level)
based on the fields that are required to apply the where clause queries
further on.