Encapsulate resultFuture in BatchIteratorBackpressureExecutor
Before the `executionFuture` could be completed from two components,
either the component using the `BatchIteratorBackpressureExecutor` or from
the `BatchIteratorBackpressureExecutor` itself. Now the
`BatchIteratorBackpressureExecutor` owns the future.

This also changes the kill handling a bit. Now we rely on
`BatchIterator.moveNext` failing if it's killed.
In addition. The `ShardingUpsertExecutor` will no longer short-circuit
but wait for pending operations to complete before setting a failure.