Allow ``!= ANY()`` to operate on arrays with > 1024 elements.
In order to use more than 1024 clauses in a Lucene BooleanQuery the
setMaxClauseCount() must be set globally (static variable). ES already
does this in the initialization of SearchModule, so we just override the
default value of the corresponding setting "indices.query.bool.max_clause_count"
in our SQLPlugin and expose this setting in our documentation.