LongType: Improve BytesRef -> long conversion error msg
The error message of a failed BytesRef -> long conversion only contained
the value of the string that couldn't be converted.

This wasn't an issue in most cases as we wrap a `NumberFormatException`
into a a different exception with a good error message if it's part of a
cast. So something like `select 'foo'::long` raised a meaningful error,
but some other cases miss this conversion.