Make template schema update synchronous
There was a window in which it was possible to create the same column
with different types in two or more partitons. Selecting from a
partitioned table with such partitions would then no longer be possible
as a column type must be consistent across partitons.

This fixes the race condition by updating the template synchronous on
the dynamic mapping update, similar how the partition/index is updated.