Introduce SingleColumnTableType as DataType of SelectSymbol
In order to support selection of composite types in subqueries and to to allow
subqueries in ANY or IN clauses, we need a wrapper type for these queries.

This models the SingleColumTableType as a CollectionType which is now the base
for other Collection-based types like ArrayType and SetType.

We remove the SingleRowSingleValueConsumer in favor of a the FirstColumnConsumer
which materializes the entire first column. For now, the SingleValueFunction
ensures that we don't return more than one value.