fixes a single insert only related memory leak
the AutoCreateIndex class is expected to be used as a singleton, but it
wasn’t and a new instance was created for every single insert.
the real memory leak was happening due to that this class is registering
itself as a setting update consumer, so every insert registered a new
instance as a settings update consumer and it was never de-registering.
also usage of AutoCreateIndex was wrong anyway, replace it with own correct logic.