Add initial semi-join rewrite for `op ANY sub-select`
For now this is a bit limited. The rewrite is only done if:

 - The root relation is a simple select from a single table
 - Only one ANY or IN expression is in the WHERE clause
 - The column used in the ANY or IN expression must also appear in the
 select list of the root relation

Furthermore, the performance will actually decrease significantly for a
lot of cases. The rewrite is only results in a better execution
strategy if the subquery resultset is huge and would lead to memory
issues with a two-phase execution.