Support generic elements in BatchIterator
This replaces `rowData() -> Columns` on the `BatchIterator` with
`currentElement() -> T`

This makes the `BatchIterator` more flexible. It'll allow us to model
proper data transformation pipelines. This could be useful in places
where we currently create batch requests.

Furthermore, the rather confusing indirections (`Columns`,
`RowBridging`, `ProxyInput`, etc..) are gone. Due to that, the
`composite` implementations are now a simpler.

The disadvantage is that it's less clear that the `BatchIterator` may
contain instances which are shared.