Run Checkstyle plugin on compile and adjust code style
This runs the Checkstyle plugin after the 'compileJava' Gradle task. The rules
are located in 'gradle/checkstyle/rules.xml'. The Checkstyle configuration is
located in 'gradle/javaModule.gradle'. The rules resemble a slightly less strict
version of the Google Java code style.

The code has been adapted to comply with the rules. For now, the changes only
affect the main source scope (no testing code). IntelliJ and other IDEs have
plugins to integrate the checks while coding which is quite useful.