Fold constant expressions early in ExpressionAnalyzer
This folds deterministic function calls which only contain literals already in
the `ExpressionAnalyzer`. This is a preparation for an overhaul of our type
precedence system. We want to normalize constant expressions early on in the
`ExpressionAnalyzer` to be able to return placeholder values which can return
various types depending on the type desired by the enclosing expression. This
enables us to simplify the type system (avoid casts) while keeping accuracy of
the folded values.