Add LogicalPlan/LogicalPlanner for aggregates on base user tables
This adds a new `LogicalPlan` and `LogicalPlanner` which should
eventually replace the `ConsumingPlanner`.

See the class documentation for a description of the `LogicalPlan` and
`LogicalPlanner`.

The goal here is to reduce the complexity that we currently have in the
`ConsumingPlanner` due to a lot of repeating logic. Almost all consumers
currently deal with limit and order by on their own. The consumers are
not really composable.

Although this commit already introduces a lot of operators which will be
useable for all types of queries, it is only enabled for aggregations on
base user table because functionality required for other queriees is
still missing or incomplete.