Reuse the MergePhase ramAccountingContext in the RoutedCollectPhase.
The merge phase creates a RamAccountingContext and a ProjectingBatchConsumer
that will use it to track the consumed rows.
In case the upstream is on the same mode, during the merge phase, we do not
register any execution context using the newly created ProjectingBatchConsumer,
but we register the consumer with the PreparerContext for a later phase to pick
it up and use it as part of an execution context.

In the case of the RoutedCollectPhase, a new RamAccountingContext was created and
that *new* context was passed in the execution context for tracking (the execution
context would close the RamAccountingContext during the cleanup phase which is
essential in order to decrement the tracked bytes in the underlying circuit breaker),
while the RamAccountingContext created in the merge phase would've been used in the
ProjectingBatchConsumer but never closed (meaning the accummulated bytes would've never
been substracted in the underlying ciruit breaker).

This commit has the MergePhase register the RamAccoutingContext it creates along with
the ProjectingBatchConsumer in the PreparerContext and the RoutedCollectPhase will use
both if present (in case no RamAccountingContext is found in the PreparerContext, a
new one is created)

(cherry picked from commit a0345d9773a3dfe7ccfc86bae605b22914591052)
Signed-off-by: Andrei Dan <andrei@crate.io>