Introduce type precedence and type inference
Type precedence during casting enables us to decide if parameters in expressions
can be safely cast to another parameter's type. This is possible by assigning
IDs to all the types which denote the rank of the types. Higher IDs denote
higher precedence.

All the casts in the ExpressionAnalyzer now follow the same
pattern: When looking up a function, we use the FunctionResolver interface which
may return new input types through the getSignature(List<DataType>) method. The
analyzer can then cast the current arguments to match the types to the
FunctionImplementation.

The type matching works through the FuncParams class. Each FuncParams is
composed of zero, one, or multiple Params which contain the valid types for the
parameters. A Param can be passed to FuncParams multiple types to connect
parameters with each other. For example, FuncParams.of(a, a, b) means that a is
of the same type. This information allows us to improve type inference.

The type inference works as follows. First, a Param is unbound. When arguments
are passed, it becomes bound. When it is bound, it will try to convert an input
to its current type, the input type, or one of the valid types provided for the
Param. For example, if we first bind to Integer and we get a Long, we convert to
Long; if we then get a String and String is a valid type, we will convert to
String.

Due to these type handling improvements, we can now avoid column casting. We
cast Literal values but require explicit casts whenever two columns are compared
which don't have a matching type. Column casts should generally be avoided
because they can slow down Lucene queries. Type precedence always has preference
when converting types but may be omitted if conversion can still be
performed (i.e. column has lower precedence).