Remove ClusterService from TableInfo instances
This changes the `getRouting` signature to also require a `ClusterState` and
`OperationRouting` and in turn removes the `ClusterService` dependency
from all `TableInfo` instances.

This is a first step to make `getRouting` deterministic. Currently we've
some complicated logic in the `RoutingBuilder` and in
`Routing.mergeLocations` to "normalize" the used shards across
relations in a join. By making `getRouting` deterministic we'll be able
to remove that. Furthermore, externalizing the `state` dependency from
the TableInfo is one more step towards a re-useable Analysis.