Allow array to geo_point conversions
In expressions like `distance(geo_point_col, [1,0])` we don't allow conversions
of columns (e.g. geo_point_col) anymore; instead, we convert the
Literal (e.g. [1,0]) to the column type (e.g. geo_point). This is currently not
possible, so we need to add another rule to allow the conversion.