Always visit fields to ensure join condition fields are resolved
The issue occurs with the following query:

select id1 from (select id1, id2, id3, id4 from j1) a
left join j2 b on a.id2 = b.id
left join j3 c on a.id3 = c.id
left join j4 d on a.id4 = d.id;

`id4` is represented as `join.a.b.a['id4']`, so a check for the relation `a`
fails in the if block which determines whether to process the field. The names
are rewritten later in the process of joining the relations.