Fixed an issue that prevented the complete graceful shutdown of CrateDB node.
The onResponse() callbacks were executed by a thread from ClusterService's threadpool.
This thread was calling `System.exit(0)` so the threadpool of ClusterServices could not shutdown properly on the first try.
Then we forced the shutdown of the threadpool, so the thread, which has called System.exit(0) and waits for the shutdown hooks to complete,
is killed and the whole process is killed before gracefully shutting down all services.

This is solved by dispatching the onResponse() work to a separate thread defined inside DecommissioningService.