add password property to CREATE USER statement
This will extend the CREATE USER statement with a password property. It allows
to specify a password for the user that is being created.

Usage:

  CREATE USER foo WITH (password = 'pass');

The password information is computed using a salted PBKDF2 algorithm that
produces a SHA-512 hash. The SecureHash (containing iterations, salt and
SHA-512 hash) is stored in the custom metdata.