Add ParameterDescription support for INSERT INTO .. VALUES
Follows the same pattern as in https://github.com/crate/crate/pull/6499

This will probably need some more modifications once we start using
the unbound analysis in the planner/executor as it lacks generated
column handling. But it's sufficient for the ParameterDescription
purposes.

`INSERT INTO [..] query` handling will be done in a separate commit/PR.