removes no-op optimization on select plans
A no-op plan was returned by the select planner if the limit of a select
resulted in 0.
All collect source are already handling where clause no-match cases, so
creating and executing a select plan instead is fine.
Turns out this optimization is not worth the increased complexity
specially when doing further refactoring of plans into logical plan
operators.