Convert timestamp to number for arithmetics
Whenever a timestamp is used in arithmetics, e.g. ts + 1 or ts / 1.0, we
previously operated in the timestamp domain. This led to problems because we
converted long/double to timestamp. Longs are converted by interpreting them as
milliseconds, whereas Doubles are converted as seconds with milliseconds as
fraction. For example:

  1.0 would be interpreted as 1000
  1   would be interpreted as 1

The solution is to treat the timestamp as long and cast it to other data types
when necessary. The result of an arithmetics with a timestamp, will never be a
timestamp but a number data type.  Only when functions accept timestamps as the
direct argument, the input will be attempted to be converted to a timestamp
and Doubles are handled special.