expose cgroup stats on sys.nodes
This commit adds cgroup cpu stats to the `os` field of sys.nodes
for Linux. The cgroup information is useful if CrateDB runs in a Docker
environment in order to provide the limits that have been provided on
the container instead of printing the host limits.

On OS other than Linux this fields will return `null`.