Integrate scalar subquery execution into ExecutionPlan building
So far we've had multiple steps to create a plan and execute them:

 - Logical planning -> returned a Plan
 - Execution planning -> (binding params) -> returned a ExecutionPlan
 - Task building -> returned a Task which was then executed

To support scalar subqueries we:

 - Built the ExecutionPlan upfront (with SelectSymbols in place)
 - During Task building we processed the dependencies
 (sub-executionPlans) first, and replaced the SelectSymbols with
 Literals

This worked so far, but it has the limitation that we cannot utilize
concrete values in all cases to build optimized execution plans.

One example is `delete from t where partition_col > (select ...)`.  The
best possible execution-plan is a plan which deletes the partitions.
But without having the concrete values and evaluating the query for all
known partitions it is not possible to create such a plan - we had to
fallback to a slower `delete-by-query` plan.

This commit overhauls the plan creation and execution:

 - Logical planning -> returns a Plan
 - Plan is executed

How the plan is executed is encapsulated in the plan.  This allows us to
have different approaches. If there are subqueries we can execute them
first and built the executionPlan incrementally using the values
returned from the subqueries. In the case of DDL statements we could
avoid both the ExecutionPlan and the Task step.

This allows us to optimize the aforementioned case and should also help
further down the road when we begin to support correlated subqueries.

As a consequence of this we might not always be able to display the full
execution plan when using the `EXPLAIN` statement.