Support subscripts on values of information_schema.table_partitions
To support something like `select values['p'] from
information_schema.table_partitions`

If the column doesn't exist it will return null instead of failing. In
addition there is also a implicit cast to string. This is necessary
because for streaming (via pg-wire); It would fail with mixed type which
could be the case if 2+ tables have the same column name for
partitioning but different types.