"Push down" UNION OrderBy to source relations
Queries like `Select id from t1 UNION ALL Select id2 from t2 order by id` are
optimized to perform an order by already in the source relations of the
union. The OrderBy from the union is "pushed down" to the queries on t1 and
t2. The result only needs to run a sorted merge.

Additionally, if two Order are discovered in a row, they will be combined.

       *Order*                   Union                  Union
          |                     /     \                /     \
          |                    /       \              /       \
        Union              *Order*  *Order*       *Order*  *Order*
       /     \                |        |             |        |
      /       \      =>       |        |     =>      |        |
   Collect   Order          Collect  Order        Collect  Collect
     t1        |              t1       |            t1       t2
               |                       |
            Collect                 Collect
              t2                      t2

Operators allowed to push down an Order: Union, RelationBoundary, FetchOrEval.

This moves the `tryCollapse` inside the newly introduced `tryOptimize` method
which traverses the entire plan. If a pushdown operator is discovered, it will
be pushed and then the rest of the graph will be traversed.