"Push down" UNION OrderBy to source relations
Queries like `Select id from t1 UNION ALL Select id2 from t2 order by id` are
optimized to perform an order by already in the source relations of the
union. The OrderBy from the union is "pushed down" to the queries on t1 and
t2. The result only needs to run a sorted merge.

       *Order*                      Union
          |                        /     \
          |                       /       \
        Union                 *Order*  *Order*
       /     \                   |        |
      /       \        =>        |        |
   Collect   Order             Collect  Order
     t1        |                 t1       |
               |                          |
            Collect                    Collect
              t2                         t2