Change subqueries with pk filters to have realtime semantics
This removes the ESGetTask and instead adds a PKLookupPhase & Operation
which ties into our execution infrastructure.

This way the "Get" specific code no longer has to take care of ORDER BY,
LIMIT, OFFSET. This is handled by other operators in the planner, and by
projections in the executor.

Furthermore it eliminates a couple of restrictions regarding the use of
`_version` in the where clause.

From a performance perspective this change seems to be an improvement as
well. Running
https://github.com/crate/crate-benchmarks/blob/master/py/specs/select.toml
yielded:

    Q: select * from uservisits where "sourceIP" = '25.193.131.52'
    C: 1
    76.82% mean difference. Likely significant
                V1  →     V2
    mean:    0.944 →   0.420
    stdev:   0.554 →   0.265
    max:     7.649 →   5.848
    min:     0.372 →   0.230