Fixed an error thrown when selecting `thread_pools['name'].
String was used instead of BytesRef.
Since integration tests don't catch this error, due to transformation happening to facilitate easy assertions,
added an assertion on the abstract class that handles those expressions to catch such errors in the future
and without adding and perf overhead for normal runtime.

Also changed the test utilities that do this transformation to throw error if they detect Strings.
This will catch errors when integ tests run without JDBC which takes place in a random manner.