Fix Paging issues caused by non-sequential bucket id
The UNION implementation introduced a MergePhase with two different input
sources. To be able to receive data from two different inputs, the bucket ids
had to contain an additional input id in addition to the node id. As a
consequence, bucket ids may not be sequential anymore. The PageDownstreamContext
relied on the sequential nature of the bucket ids and had to be adjusted.

This commit introduces a Set to hold all existing input ids. While sequential
ids allowed to loop over all input ids without storing the input ids itself, the
Set now replaces this functionality by looping over it to retrieve more data
from the inputs.

This fixes issues where the loading of more paging data was triggered
incorrectly because the number of finished buckets was not accounted for
correctly. The locking in the PageDownstreamContext was improved and the old
logic which sets buckets to empty was re-added. A test was added to test
processing of the previously added non-sequential bucket ids.

Regression of

1bf8efe5555087390436a01e26ca6ed49fe9b24d
(initial introduction of non-sequential bucket ids)

dd4ba3a436f93e68ec6e8f267e9a96c305bdc017
(fix to re-create buckets which have been cleared)