Move benchmarks into a separate module
Some benchmarks didn't run anymore because `Randomness.get()` causes an
error if the `randomizedtesting` dependencies are in the classpath, but
the code in question is called *without* a RandomizedContext being
setup. This setup is usually done by the TestRunner, but jmh doesn't run
in a TestRunner, and it's not recommended to do that, as it would
influence the benchmark results.

This introduces a new benchmark module. This allows us to remove the
`randomizedrunner` dependencies from the benchmarks.

This also removes the `jmh` plugin in favor of simply using a jmh jar.
This has the advantage that tests can be filtered more conveniently.
(All jmh CLI arguments can be used)