Fixed incorrect pushdown of LIMIT in case of > 2 table joins.
When more than 2 tables are involved, we pushdown the limit to
all TwoTableJoin instances of the join tree. Then we detect if there
is filtering involved in the join tree and we remove any pushed-down
limit before the last filtering. Before the fix we detected as filtering
only the filtering happening in where clauses but not the filtering happening
because of the ``ON <condition>`` join conditions.

Fixes: https://github.com/crate/crate/issues/6641