Use a collector dispatcher when a shard is not available for collection.
When we attempt to collect the rows we might run into situations where the
shards are temporary or permanently unavailable (ie table dropped).
In the later case, we fail the operation.
In the former case however, instead of failing, we now monitor the target
shard's state and resume collection when it becomes available again (this
can translate into a remote collection in case the shard relocated to a
different node or a local collection if the shard is still/back on the same
node)