re-analyze where clause after sub-query excution
Fixes an error if a subquery is used in comparison with a routing/clustered-by column.

This fix requires to adjust how an explain plan output is generated because
a logical plan containing a sub-query symbol cannot be built into an execution plan without
executing the sub-queries. Introducing a logical plan printer which kicks in
if building an execution plan (and so using the execution plan printer) fails.