re-analyze where clause after sub-query excution
This fixes:
 - that a query was executed as a table scan and a generic "null" filter
   instead of immediately returning a NO-MATCH result when the included subquery
   returns no result.
 - that a exception was raised when a subquery was used in comparison with a
   routing/clustered-by column.

Also it was required to adjust how an explain plan output is generated because
a logical plan containing a sub-query symbol cannot be built into an execution plan without
executing the sub-queries. Introducing a logical plan printer which kicks in
if building an execution plan (and so using the plan printer) fails.