Optimize performance of IN or = ANY with subquery.
The IN() or = ANY() are translated to a Lucene's TermInSetQuery.
The TermInSetQuery is sorting the list of values to optimize the execution
of the Lucene query. In this case, since those values are the results of the
subquery we can optimize the time spent for the construction of the TermInSetQuery
object by presorting those values in our subquery execution.

This small perf optimization can significantly improve total execution time
when a large number shard of the outer table is queried. That's because for
each shard a TermInSetQuery is built and therefore a sorting of the values
takes place.

Related to: https://github.com/crate/crate/issues/6755