Change NullFieldComparatorSource to not extend XFieldComparatorSource
Since we're not using the ES fetch/merge functionality we don't need a
`XFieldComparatorSource`, a regular `FieldComparatorSource` is
sufficient.

This also makes upgrading to ES 6 a bit easier as the
`XFieldComparatorSource` is going to change a bit.

In addition this commit also changes the `missingValue` implementation
to match the implementation of `XFieldComparatorSource#missingObject`