fixes NULL handling while normalizing the WHERE clause
If an argument of a logical operator evaluated to null, this operator
was not normalized and ended up at the lucene query builder which isn’t
able of converting this type of query into a lucene query and so created
a generic function filter.
This introduces a logic to replace NULL’s with FALSE inside a where
clause which results in a normalizable query whenever possible.