Fix array assignments and type inference of params inside arrays
Assignments like `SET ob = [?]` where `ob` is an object array didn't
work correctly because a `to_object_array` cast is not supported and the
type of `[?]` was inferred as `array(undefined)`.

Instead of adding the object array cast function, this commit changes
the `cast` implementation of a `_array` `Function` symbol to propagate
to its elements. This has the advantage that type inference for the
`ParameterDescription` message will work as well.

Fixes https://github.com/crate/crate/issues/6835