Added OperationOnInaccessibleRelationException to replace UnsupportedFeatureException when dealing with closed/readonly/blocked operations
When attempting an operation on a closed table as a user, instead of the expected exception of an Unsupported Feature being raised, it would instead raise a nonsensical missing privileges exception. This is because this operation would originally throw an UnsupportedFeatureException, which would have no information about the table it was performed on. The privilege checking mechanism would then attempt to check the privileges on nothing and, finding nothing there, throw the incorrect exception.