Fix PreExecutionBenchmarks
With the ES6 upgrade we started to make use of
`buildNewFakeTransportAddress`, this is problematic because it causes
`RandomizedTest` to be loaded, resulting in `NoClassDefFoundError`:

java.lang.NoClassDefFoundError: com/carrotsearch/randomizedtesting/RandomizedTest
	at org.apache.lucene.util.LuceneTestCase.<clinit>(LuceneTestCase.java:393)
	at io.crate.testing.DiscoveryNodes.newNode(DiscoveryNodes.java:42)
	at io.crate.analyze.PreExecutionBenchmark.setup(PreExecutionBenchmark.java:74)

This is because for benchmarks we need to exclude randomizedtesting.
Otherwise we'd run into RandomizedContxt errors because benchmarks are
not run within a test runner.

This re-implements `buildNewFakeTransportAddress` to avoid that.